Start managing fuzzer message processors (WIP)
Begin plumbing in API endpoints that allow management of the
selected message processors, modeled on how options are supported.

The interesting functions are still missing, but we have now define
(and consistently use) a standard processor list when we create a
fuzzer.

Don't try to build - just saving changes so I don't lose them over the
weekend... :-O