Change "add" strategy
Make the caller supply enough information for a processor script that it
is possible to construct the required wrapper object from scratch. It is
expected that a side-effect will be that attempts to add previously-
loaded scripts will fail with duplicate name errors.

Change method names to remove any pretense of handling non-script
processors.

Extend documentation to mention that it's possible to mess with the
processor list now.